+++ odoo/songs/install/data_all.py
@@ -266,6 +266,13 @@ def desactive_incoterm(ctx):
 
 
 @anthem.log
+def import_price_category(ctx):
+    """ Importing price category from csv """
+    load_csv(ctx, 'data/install/price_category.csv',
+             'product.price.category')
+
+
+@anthem.log
 def main(ctx):
     """ Loading data """
     create_analytic_dimension(ctx)
@@ -294,3 +301,4 @@ def main(ctx):
     create_layout_category(ctx)
     import_account_asset_category(ctx)
     desactive_incoterm(ctx)
+    import_price_category(ctx)